requirements.txt
httpx>=0.26
pydantic>=2.6
PyYAML>=6.0
jsonschema>=4.21
androguard>=3.4.0
apkInspector>=1.1.7
lxml>=4.3.0
rich>=13.7
opentelemetry-api>=1.24
opentelemetry-sdk>=1.24
opentelemetry-exporter-otlp>=1.24
